requirements.txt
langchain==0.3.3
langchain_core==0.3.12
langchain_google_community==2.0.1
langchain_openai==0.2.2
langgraph==0.2.38
nest_asyncio==1.6.0
python-dotenv==1.0.1
typing_extensions==4.12.2
streamlit~=1.39.0